Fixed errant Double dependency in lift op, changed test to int to ensure it's not using DM's dv * dm, which is specialized for blas with Double.

+++ math/src/test/scala/breeze/linalg/DenseVectorTest.scala
@@ -171,11 +171,11 @@ class DenseVectorTest extends FunSuite with Checkers {
   }
 
   test("DenseVector * DenseMatrix Lifted OpMulMatrix") {
-    val x = DenseVector[Double](1, 2, 3)
-    val m = DenseMatrix((1.0,2.0,3.0),
-                        (2.0,4.0,6.0),
-                        (3.0,6.0,9.0))
-    val mr = DenseMatrix((1.0,2.0,3.0))
+    val x = DenseVector[Int](1, 2, 3)
+    val m = DenseMatrix((1,2,3),
+                        (2,4,6),
+                        (3,6,9))
+    val mr = DenseMatrix((1,2,3))
 
     val xxt = x * x.t
     assert(xxt === m)
